+++ demisto_sdk/commands/common/content/objects_factory.py
@@ -1,5 +1,7 @@
 from typing import Union
 
+from wcmatch.pathlib import Path
+
 from demisto_sdk.commands.common.constants import OLD_INDICATOR_TYPE, FileType
 from demisto_sdk.commands.common.content.objects.pack_objects import (
     AgentTool, AuthorImage, ChangeLog, Classifier, ClassifierMapper,
@@ -11,7 +13,6 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
 from demisto_sdk.commands.common.content.objects.root_objects import \
     Documentation
 from demisto_sdk.commands.common.tools import find_type
-from wcmatch.pathlib import Path
 
 from .errors import ContentFactoryError