+++ demisto_sdk/commands/format/update_description.py
@@ -2,6 +2,7 @@ import re
 from typing import Tuple
 
 import click
+
 from demisto_sdk.commands.common.hook_validations.description import \
     DescriptionValidator
 from demisto_sdk.commands.format.format_constants import (ERROR_RETURN_CODE,