+++ demisto_sdk/commands/format/update_widget.py
@@ -1,7 +1,8 @@
-from typing import Tuple
 from distutils.version import LooseVersion
+from typing import Tuple
 
 import click
+
 from demisto_sdk.commands.format.format_constants import (ERROR_RETURN_CODE,
                                                           SKIP_RETURN_CODE,
                                                           SUCCESS_RETURN_CODE)