+++ demisto_sdk/commands/openapi_codegen/openapi_codegen.py
@@ -8,6 +8,7 @@ from typing import Any, List, Optional, Union
 
 import autopep8
 import yaml
+
 from demisto_sdk.commands.common.hook_validations.docker import \
     DockerImageValidator
 from demisto_sdk.commands.common.tools import camel_to_snake, print_error
@@ -672,6 +673,9 @@ class OpenAPIIntegration:
                     for ref in refs:
                         ref = ref.split('/')[-1]
                         ref_props = self.extract_values(self.reference.get(ref, {}), 'properties')
+                        # Addition of filtering dicts only was added because some swaggers contain example files
+                        # Which are written in string and caused errors on ref_props[0].items()
+                        ref_props = [ref_prop for ref_prop in ref_props if isinstance(ref_prop, dict)]
                         if ref_props:
                             for k, prop in ref_props[0].items():
                                 if k in self.root_objects:
@@ -719,6 +723,9 @@ class OpenAPIIntegration:
                 for ref in refs:
                     ref = ref.split('/')[-1]
                     ref_args = self.extract_values(self.reference.get(ref, {}), 'properties')
+                    # Addition of filtering dicts only was added because some swaggers contain example files
+                    # Which are written in string and caused errors on ref_props[0].items()
+                    ref_args = [ref_arg for ref_arg in ref_args if isinstance(ref_arg, dict)]
                     for ref_arg in ref_args:
                         for k, v in ref_arg.items():
                             new_ref_arg = {'name': k, 'in': arg.get('in'),