+++ demisto_sdk/commands/test_content/tests/build_context_test.py
@@ -1,6 +1,7 @@
 import json
 
 import pytest
+
 from demisto_sdk.commands.test_content.ParallelLoggingManager import \
     ParallelLoggingManager
 from demisto_sdk.commands.test_content.TestContentClasses import BuildContext