+++ demisto_sdk/commands/update_release_notes/update_rn_manager.py
@@ -2,6 +2,7 @@ import os
 from typing import Optional, Tuple
 
 import git
+
 from demisto_sdk.commands.common.constants import (
     API_MODULES_PACK, SKIP_RELEASE_NOTES_FOR_TYPES)
 from demisto_sdk.commands.common.tools import (LOG_COLORS,
@@ -9,8 +10,8 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS,
                                                filter_files_on_pack,
                                                get_pack_name,
                                                get_pack_names_from_files,
-                                               print_color,
-                                               print_warning)
+                                               print_color, print_warning,
+                                               suppress_stdout)
 from demisto_sdk.commands.update_release_notes.update_rn import (
     UpdateRN, update_api_modules_dependents_rn)
 from demisto_sdk.commands.validate.validate_manager import ValidateManager
@@ -37,7 +38,7 @@ class UpdateReleaseNotesManager:
         # When a user choose a specific pack to update rn, the -g flag should not be passed
         if self.given_pack and self.is_all:
             raise ValueError('Please remove the -g flag when specifying only one pack.')
-        self.rn_path = list()
+        self.rn_path: list = list()
 
     def manage_rn_update(self):
         """
@@ -77,6 +78,12 @@ class UpdateReleaseNotesManager:
             if not validate_manager.git_util:  # in case git utils can't be initialized.
                 raise git.InvalidGitRepositoryError('unable to connect to git.')
             validate_manager.setup_git_params()
+            if self.given_pack:
+                # The Validator prints errors which are related to all changed files that were changed against prev
+                # version. When the user is giving a specific pack to update, we want to suppress the error messages
+                # which are related to other packs.
+                with suppress_stdout():
+                    return validate_manager.get_changed_files_from_git()
             return validate_manager.get_changed_files_from_git()
         except (git.InvalidGitRepositoryError, git.NoSuchPathError, FileNotFoundError) as e:
             raise FileNotFoundError(